Rework in regard to review comments
* Add type hints
* Factor out helper function(s) to keep noxfile.py clean
